Re-enable getprocesses, overlay, and keyboard watcher
The overlay does work on Linux, but with some quirks
(mainly on tiling window managers).

Luckly the user can disable the overlay if they wish.

My fork of memoryjs does support getProcesses, so I've re-enabled that
code.

I also have a linux port of 'node-keyboard-watcher' which I'd like to
optimize, so that it can be put into a pull-request.